Ajout MatchService
- Ajout gestion des matches MatchService
- Correction petite erreur de type dans TeamController

+++ TeamService/src/main/java/com/m2dfs/TeamService/controller/TeamController.java
@@ -23,7 +23,7 @@ public class TeamController {
     }
 
     @GetMapping("/{id}")
-    public Team getTeamById(@PathVariable Long id) {
+    public Team getTeamById(@PathVariable int id) {
         return teamDB.get(id);
     }
